+++ R/zzz.R
@@ -7,11 +7,17 @@
     else
       path <- tempdir()
     set_resource_path(path = path, verbose = FALSE)
-    }
+  }
+  # WHO
   makeActiveBinding(sym = "icd10who2016",
                     fun = .icd10who2016_binding,
                     env = parent.env(environment()))
   lockBinding(sym = "icd10who2016", env = parent.env(environment()))
+  makeActiveBinding(sym = "icd10who2008fr",
+                    fun = .icd10who2008fr_binding,
+                    env = parent.env(environment()))
+  lockBinding(sym = "icd10who2008fr", env = parent.env(environment()))
+  # ICD-10 CM
   if (!("icd.data.icd10cm_active_ver" %in% names(options()))) {
     set_icd10cm_active_ver(2019, check_exists = FALSE)
   }